Refactor NpcPrompt to be reusable and update derived classes
- Moved common functionality to NpcPrompt.cs in the interface folder
- Updated GarrethPrompt.cs and SilasPrompt.cs to inherit from NpcPrompt
- Changed DisplayDescription method in NpcPrompt to protected for accessibility
